requirements.txt
streamlit==1.50.0
torch==2.4.1
torchvision==0.19.1
pillow==10.3.0
